requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio-dgram==2.1.2
attrs==22.2.0
autopep8==2.0.1
bcrypt==4.0.1
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
discord.py==2.2.0
dnspython==2.2.1
frozenlist==1.3.3
idna==3.4
mcstatus==10.0.2
multidict==6.0.4
packaging==23.0
paramiko==3.0.0
pip-review==1.3.0
pkg_resources==0.0.0
pycodestyle==2.10.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
yarl==1.8.2
